Don't access / memoize request.params
Accessing request.params triggers memoization (ie. it stores the params in the RACK_REQUEST_FORM_HASH header.)

This can be problematic when used together with the roda :json_parser-plugin. I determined that using PUT here is actually unnecessary, as this is only a contract inside this gem, so we can just use POST, and get around the unwanted memoizing of params.